Exercise 1.12: Project v0.6
Get a random picture from Lorem Picsum like
https://picsum.photos/1200 and display it in
the project. Find a way to store the image
so it stays the same for 60 minutes.

Make sure to cache the image into a volume
so that the API isn't needed for new images
every time we access the application or the
container crashes.

Also:

- Install yamllint and take it into use with project templates.
- Install "Tim-Koehler.helm-intellisense" extension.