Exercise 2.02: Project v1.0
"The next step is to create a new container that
takes care of saving the todo items. This new service,
let us call it todo-backend, should have a GET /todos
endpoint for fetching the list of todos and a POST /todos
endpoint for creating a new todo. The todos can be saved
into memory, we'll add a database later.ty message aborts
the commit."

Also:
- Variable dependency management robustness improved between applications
- Added .js file linter
- Most of the templates are now universally scalable, removed redundancy
- Debugging feature to backend application that allows debugging the
  application by simply switching the container names order in values.yaml.
- Logic added that allows the application to be started in an
  order where each of them will start succesfully first time.
- Generate Chart.yaml files automatically in helm template resolver to
avoid redundant definitions